DAOS-4170 control: Simplify pool creation
The `dmg pool create` command now has a --size parameter which
can be used to specify the total size of a pool. The --scm-ratio
and --num-ranks parameters may be optionally specified if the
defaults (6% SCM:NVMe, all ranks) are not suitable.

The existing --scm-size and --nvme-size flags remain, but should
generally not be used and will probably be removed eventually.

The pool create output has also been updated to provide more
detailed information about how the pool was created.

Signed-off-by: Michael MacDonald <mjmac.macdonald@intel.com>